Stage 7: introdotti tipi SelectorDelegate e basi per static mock
STATIC MOCKING
mocking statico richiede mockito-inline sotto 4.11, quindi si preferisce
aggiornare Mockito a 5.19 che non richiede dipendenza aggiuntiva.

Tuttavia, sia sotto 4.11 (con inline) che ancora 5.19, Mockito si serve
di agent attachment. Il modo sicuro di farlo è attraverso un argument

	-javaagent ...

che quindi abbiamo
	1. aggiunto nel POM a surefire, con accortezza che
	   maven-dependency-plugin deve eseguire per rendere noto a surefire
	   il path del jar di mockito-core
	2. aggiunto negli args di una Run Configuration di Eclipse, usando
	   per ora un path hard al jar di Mockito

a questo punto, una JVM OpenJDK 21 emette comunque uno warning ma meno
invasivo del precedente

Comunque, CONSIDERA: agent attachment sembra rallentare i test in
Eclipse

TIPI SELECTOR DELEGATE
Servono a LineUpChooser per poter fare assunzioni circa le capacità dei
Selector restituiti dai Delegate di primo livello, ovvero: invocare
creazione di un Dealing Group e di una Sequence sui selettori

+++ fantacalcio-app/src/main/java/businessLogic/UserService.java
@@ -5,72 +5,71 @@ import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 
+import businessLogic.repositories.LeagueRepository;
 import domainModel.*;
 
 public class UserService {
 
-	private final TransactionManager transactionManager;
-
+	protected final TransactionManager transactionManager;
+	
 	public UserService(TransactionManager transactionManager) {
 		this.transactionManager = transactionManager;
 	}
-
+	
 	// League
-	public Optional<League> existingLeague(String leagueCode) {
-		return transactionManager.fromTransaction(
-				(context) -> context.getLeagueRepository().getLeagueByCode(leagueCode));
-	}
 
-	public Set<League> getLeaguesByUser(FantaUser user) {
-		return transactionManager.fromTransaction((context) ->
-			context.getLeagueRepository().getLeaguesByUser(user));
-
-	}
-
-	public void createLeague(String leagueName, FantaUser fantaUser, NewsPaper newsPaper, String leagueCode) {
+	public void joinLeague(FantaTeam fantaTeam, League league) {
 		transactionManager.inTransaction((context) -> {
-			if(existingLeague(leagueCode).isEmpty()) {
-				League league = new League(fantaUser, leagueName, newsPaper, leagueCode);
-				context.getLeagueRepository().addLeague(league);
-			}else{
-				throw new IllegalArgumentException("A league with the same league code already exists");
+
+			LeagueRepository leagueRepository = context.getLeagueRepository();
+
+			final int maxFantaTeamsPerLeague = 12;
+			if (leagueRepository.getAllTeams(league).size() >= maxFantaTeamsPerLeague)
+				throw new UnsupportedOperationException("Maximum 12 teams per league");
+
+			FantaUser user = fantaTeam.getFantaManager();
+			List<League> UserLeagues = leagueRepository.getLeaguesByUser(user);
+			if (UserLeagues.contains(league)) {
+				throw new IllegalArgumentException("You have already a team in this league");
+			} else {
+				context.getTeamRepository().saveTeam(fantaTeam);
 			}
 		});
 	}
 
-	public void joinLeague(FantaTeam fantaTeam, League league) {
-		 transactionManager.inTransaction((context) -> {
-			 FantaUser user = fantaTeam.getFantaManager();
-			 Set<League> UserLeagues = context.getLeagueRepository().getLeaguesByUser(user);
-			 if(UserLeagues.contains(league)) {
-				 throw new IllegalArgumentException("You have already a team in this league");
-			 } else {
-				 context.getTeamRepository().saveTeam(fantaTeam);
-			 }
-		 });
-	}
-
 	// Matches
 
-	public Map<MatchDaySerieA, Set<Match>> getAllMatches(League league) {
-		return transactionManager.fromTransaction(
-				(context) -> context.getMatchRepository().getAllMatches(league));
+	public Map<MatchDaySerieA, List<Match>> getAllMatches(League league) {
+
+		return transactionManager.fromTransaction((context) -> {
+
+			List<MatchDaySerieA> allMatchDays = context.getMatchDayRepository().getAllMatchDays();
+			Map<MatchDaySerieA, List<Match>> map = new HashMap<MatchDaySerieA, List<Match>>();
+
+			for (MatchDaySerieA matchDay : allMatchDays) {
+				map.put(matchDay, context.getMatchRepository().getAllMatchesByMatchDay(matchDay, league));
+			}
+
+			return map;
+
+		});
 	}
 
 	public Match getNextMatch(League league, FantaTeam fantaTeam, LocalDate date) {
 		return transactionManager.fromTransaction((context) -> {
 			Optional<MatchDaySerieA> previousMatchDay = context.getMatchDayRepository().getPreviousMatchDay(date);
 			if (previousMatchDay.isPresent()) {
-				Match previousMatch = context.getMatchRepository().getMatchByMatchDay(previousMatchDay.get(), league, fantaTeam);
+				Match previousMatch = context.getMatchRepository().getMatchByMatchDay(previousMatchDay.get(), league,
+						fantaTeam);
 				Optional<Result> result = context.getResultsRepository().getResult(previousMatch);
-				if(result.isEmpty()){
+				if (result.isEmpty()) {
 					throw new RuntimeException("The results for the previous match have not been calculated yet");
 				}
 			}
 			Optional<MatchDaySerieA> nextMatchDay = context.getMatchDayRepository().getNextMatchDay(date);
-			if(nextMatchDay.isEmpty()){
+			if (nextMatchDay.isEmpty()) {
 				throw new RuntimeException("The league ended");
-			}else{
+			} else {
 				return context.getMatchRepository().getMatchByMatchDay(nextMatchDay.get(), league, fantaTeam);
 			}
 		});
@@ -79,46 +78,61 @@ public class UserService {
 	// Players
 
 	public List<Player> getAllPlayers() {
-		return transactionManager.fromTransaction(
-				(context) -> context.getPlayerRepository().findAll());
+		return transactionManager.fromTransaction((context) -> context.getPlayerRepository().findAll());
 	}
 
 	public List<Player> getPlayersBySurname(String surname) {
-		return transactionManager.fromTransaction(
-				(context) -> context.getPlayerRepository().findBySurname(surname));
+		return transactionManager.fromTransaction((context) -> context.getPlayerRepository().findBySurname(surname));
 	}
-
 
 	// Proposals
 
 	public List<Proposal> getAllTeamProposals(League league, FantaTeam team) {
-		return transactionManager.fromTransaction(
-				(context) -> context.getProposalRepository().getMyProposals(league, team));
+		return transactionManager
+				.fromTransaction((context) -> context.getProposalRepository().getMyProposals(league, team));
 	}
 
-	public void acceptProposal(Proposal.PendingProposal proposal) {
-		transactionManager.inTransaction(
-				(context) -> {
-					Contract receivedContract = new Contract(proposal.getOfferedContract().getTeam(),
-							proposal.getRequestedContract().getPlayer());
-					Contract givenContract = new Contract(proposal.getRequestedContract().getTeam(),
-							proposal.getOfferedContract().getPlayer());
-					context.getContractRepository().deleteContract(proposal.getRequestedContract());
-					context.getContractRepository().deleteContract(proposal.getOfferedContract());
-					context.getContractRepository().saveContract(receivedContract);
-					context.getContractRepository().saveContract(givenContract);
-					context.getProposalRepository().deleteProposal(proposal);
-				});
+	public void acceptProposal(Proposal.PendingProposal proposal, FantaTeam fantaTeam) {
+		transactionManager.inTransaction((context) -> {
+			FantaTeam requestingTeam = proposal.getRequestedContract().getTeam();
+			FantaTeam offeringTeam = proposal.getOfferedContract().getTeam();
+			Player offeredPlayer = proposal.getOfferedContract().getPlayer();
+			Player requestedPlayer = proposal.getRequestedContract().getPlayer();
+			if (!requestingTeam.equals(fantaTeam)) {
+				throw new IllegalArgumentException("You are not involved in this proposal");
+			}
+			Optional<Contract> requestedContract = searchContract(fantaTeam, requestedPlayer);
+			Optional<Contract> offeredContract = searchContract(offeringTeam, offeredPlayer);
+
+			if (requestedContract.isEmpty() || offeredContract.isEmpty()) {
+				rejectProposal(proposal, fantaTeam);
+				throw new IllegalArgumentException("One or both players do not play anymore in the teams");
+			}
+			Contract receivedContract = new Contract(proposal.getOfferedContract().getTeam(),
+					proposal.getRequestedContract().getPlayer());
+			Contract givenContract = new Contract(proposal.getRequestedContract().getTeam(),
+					proposal.getOfferedContract().getPlayer());
+			context.getContractRepository().deleteContract(proposal.getRequestedContract());
+			context.getContractRepository().deleteContract(proposal.getOfferedContract());
+			context.getContractRepository().saveContract(receivedContract);
+			context.getContractRepository().saveContract(givenContract);
+			context.getProposalRepository().deleteProposal(proposal);
+		});
 	}
 
-	public void rejectProposal(Proposal.PendingProposal proposal) {
-		transactionManager.inTransaction(
-				(context) -> {
-					Proposal.RejectedProposal rejectedProposal = new Proposal.
-							RejectedProposal(proposal.getOfferedContract(), proposal.getRequestedContract());
-					context.getProposalRepository().deleteProposal(proposal);
-					context.getProposalRepository().saveProposal(rejectedProposal);
-				});
+	public void rejectProposal(Proposal.PendingProposal proposal, FantaTeam fantaTeam) {
+		transactionManager.inTransaction((context) -> {
+			FantaTeam requestingTeam = proposal.getRequestedContract().getTeam();
+			FantaTeam offeringTeam = proposal.getOfferedContract().getTeam();
+
+			if (!requestingTeam.equals(fantaTeam) && !offeringTeam.equals(fantaTeam)) {
+				throw new IllegalArgumentException("You are not involved in this proposal");
+			}
+			Proposal.RejectedProposal rejectedProposal = new Proposal.RejectedProposal(proposal.getOfferedContract(),
+					proposal.getRequestedContract());
+			context.getProposalRepository().deleteProposal(proposal);
+			context.getProposalRepository().saveProposal(rejectedProposal);
+		});
 	}
 
 	public boolean createProposal(Player requestedPlayer, Player offeredPlayer, FantaTeam myTeam,
@@ -128,18 +142,18 @@ public class UserService {
 		}
 
 		return transactionManager.fromTransaction((context) -> {
-			Optional<Contract> requestedContract = opponentTeam.getContracts().stream().filter(c -> c.getTeam().equals(opponentTeam) && c.getPlayer().equals(requestedPlayer)).findFirst();
-			Optional<Contract> offeredContract = myTeam.getContracts().stream()
-							.filter(c -> c.getTeam().equals(myTeam) && c.getPlayer().equals(offeredPlayer))
-							.findFirst();
-			if(requestedContract.isPresent() && offeredContract.isPresent()){
+			Optional<Contract> requestedContract = searchContract(opponentTeam, requestedPlayer);
+			Optional<Contract> offeredContract = searchContract(myTeam, offeredPlayer);
+
+			if (requestedContract.isPresent() && offeredContract.isPresent()) {
 				Proposal newProposal = new Proposal.PendingProposal(offeredContract.get(), requestedContract.get());
 
-				if (context.getProposalRepository().proposalExists(newProposal)) {
+				if (context.getProposalRepository().getProposal(offeredContract.get(), requestedContract.get())
+						.isPresent()) {
 					throw new IllegalArgumentException("The proposal already exists");
 				}
 				return context.getProposalRepository().saveProposal(newProposal);
-			} else{
+			} else {
 				return false;
 			}
 		});
@@ -155,51 +169,63 @@ public class UserService {
 	// Teams
 
 	public List<FantaTeam> getAllFantaTeams(League league) {
-		return transactionManager.fromTransaction(
-				(context) -> context.getTeamRepository().getAllTeams(league));
+		return transactionManager.fromTransaction((context) -> context.getTeamRepository().getAllTeams(league));
 	}
 
 	public FantaTeam getFantaTeamByUserAndLeague(League league, FantaUser user) {
-		return transactionManager.fromTransaction((context)->
-				context.getTeamRepository().getFantaTeamByUserAndLeague(league, user));
+		return transactionManager
+				.fromTransaction((context) -> context.getTeamRepository().getFantaTeamByUserAndLeague(league, user));
 	}
 
 	// Grades
 
-	public List<Grade> getAllMatchGrades(League league, Match match) {
-		return transactionManager.fromTransaction(
-				(context) -> context.getGradeRepository().getAllMatchGrades(match, league));
+	public List<Grade> getAllMatchGrades(Match match, NewsPaper newsPaper) {
+		return transactionManager
+				.fromTransaction((context) -> context.getGradeRepository().getAllMatchGrades(match, newsPaper));
 	}
 
-	//Results
+	// Results
 
-	public Optional<Result> getResultByMatch(League league, Match match) {
-		return transactionManager.fromTransaction((context)-> context.getResultsRepository().getResult(match));
+	public Optional<Result> getResultByMatch(Match match) {
+		return transactionManager.fromTransaction((context) -> context.getResultsRepository().getResult(match));
 
 	}
-	//TODO spostarlo nell'admin service
-// va nell'admin
-/*	public void saveResult(Result result) {
-		transactionManager.inTransaction((context) -> context.getResultsRepository().saveResult(result));
-	}
-*/
-	public Optional<LineUp> getLineUpByMatch(League league, Match match, FantaTeam fantaTeam) {
-		return transactionManager.fromTransaction((context) ->
-				context.getLineUpRepository().getLineUpByMatchAndTeam(league, match, fantaTeam));
-    }
 
-	public void saveLineUp(LineUp lineUp){
-		transactionManager.inTransaction((context)-> {
-			LocalDate today = LocalDate.now();
+	public Optional<LineUp> getLineUpByMatch(Match match, FantaTeam fantaTeam) {
+		return transactionManager
+				.fromTransaction((context) -> context.getLineUpRepository().getLineUpByMatchAndTeam(match, fantaTeam));
+	}
+
+	protected LocalDate today() {
+		return LocalDate.now();
+	}
+
+	public void saveLineUp(LineUp lineUp) {
+		transactionManager.inTransaction((context) -> {
+			LocalDate today = today();
 			DayOfWeek day = today.getDayOfWeek();
-			
+
+			// Check if is a valid date to save the lineUp
 			Match match = lineUp.getMatch();
-			if(today.isAfter(match.getMatchDaySerieA().getDate()))
+			LocalDate matchDate = match.getMatchDaySerieA().getDate();
+			if (today.isAfter(matchDate))
 				throw new UnsupportedOperationException("Can't modify the lineup after the match is over");
-			
-			if(day == DayOfWeek.SATURDAY || day == DayOfWeek.SUNDAY)
+
+			if (day == DayOfWeek.SATURDAY || day == DayOfWeek.SUNDAY)
 				throw new UnsupportedOperationException("Can't modify the lineup during Saturday and Sunday");
-			
+
+			// Check if is legal to save the lineUP
+			Optional<MatchDaySerieA> previousMatchDay = context.getMatchDayRepository().getPreviousMatchDay(matchDate);
+			if (previousMatchDay.isPresent()) {
+				Match previousMatch = context.getMatchRepository().getMatchByMatchDay(previousMatchDay.get(),
+						lineUp.getTeam().getLeague(), lineUp.getTeam());
+				Optional<Result> previousMatchResult = context.getResultsRepository().getResult(previousMatch);
+
+				if (previousMatchResult.isEmpty()) {
+					throw new UnsupportedOperationException("The grades for the previous match were not calculated");
+				}
+			}
+
 			League league = lineUp.getTeam().getLeague();
 			FantaUser user = lineUp.getTeam().getFantaManager();
 			FantaTeam team = getFantaTeamByUserAndLeague(league, user);
@@ -225,13 +251,20 @@ public class UserService {
 			// Validate ownership
 			for (Player player : allPlayers) {
 				if (!teamPlayers.contains(player)) {
-					throw new IllegalArgumentException("Player " + player + " does not belong to FantaTeam " + team.getName());
+					throw new IllegalArgumentException(
+							"Player " + player + " does not belong to FantaTeam " + team.getName());
 				}
 			}
-			
-			if(getLineUpByMatch(team.getLeague(), match, team).isPresent())
+
+			if (getLineUpByMatch(match, team).isPresent())
 				context.getLineUpRepository().deleteLineUp(lineUp);
-			
-			context.getLineUpRepository().saveLineUp(lineUp);});
+
+			context.getLineUpRepository().saveLineUp(lineUp);
+		});
+	}
+
+	protected Optional<Contract> searchContract(FantaTeam team, Player player) {
+		return team.getContracts().stream().filter(c -> c.getTeam().equals(team) && c.getPlayer().equals(player))
+				.findFirst();
 	}
 }
